removed signaling on User delete
User Storage will just have to be deleted with a call to the model's
delete() method, when the user clicks the delete button.

+++ capstoneproject/models/models/word.py
@@ -43,15 +43,17 @@ class Word(Model):
         return not self.default
 
     def isRelated(self):
-        return len(self.user_storage.all()) > 0 or len(self.word_features.all()) > 0
+        return len(self.user_storage.all()) > 0 \
+            or len(self.word_features.all()) > 0
 
     def isOrphaned(self):
-        return len(self.user_storage.all()) == 0 and len(self.word_features.all()) == 0
+        return len(self.user_storage.all()) == 0 \
+            and len(self.word_features.all()) == 0
 
     def delete_relatives(self):
         word_features = list(self.word_features.all())
         for word_feature in word_features:
-            if word_feature.isOrphaned():
+            if word_feature.isOrphaned() and word_feature.isCustom():
                 word_feature.delete()
 
     def get_word_features(self):
